Fixed the wrong carrier name displayed on the UI
The spn from EF_SPN on LGE U+ SIM is different
from the plmn. Thus the UI shows "LG U+|LGU+"
as the carrier name on the UI. The fix is actually
a workaround to override the spn with "LG U+" which is
same as plmn name. The duplicate detection mechanism
will automatically rip off the spn from the name
displayed on the UI.

bug: 23506080
Change-Id: I264ba8ff8b5dcecf09598335694f69b50d6b463c