Adding 111,113,117,122,125 to the list of emergency numbers.
These are emergency numbers for Korea. Need to be added separately
for the scenario where SIM is absent and radio is off (like APM).
Default list used in that case only includes 911, 112 and a limited set
of emergency numbers, and other numbers are allowed only when sim is
available or radio is on.

Bug: 23657879
Change-Id: I494c506ed6b285ab83bfac1d790a51cbed3e8e00