Remove useImsAlwaysForEmergencyCall, it has been moved to CarrierConfig
This was set to false on mnc-dr-dev as a temporary fix. However that
does not work on master since it has been moved to CarrierConfig.
Removing this for now to fix the build. If we decide to keep it as a 
permanent fix/workaround, will need to add it in CarrierConfig overlay 
on master.

Bug: 24499450

Change-Id: I496926f49fdd1f402c19efefcaa43e451b59a767