dontaudit perfd self:capability kill;
perfd sends a kill(PID, 0) (a "null" kill) to determine if a process
exists. A process exists if the kill succeeds, or it returns with
errno=EPERM. A process doesn't exist if the kill fails with
errno=ESRCH. In either case, CAP_KILL is not needed to get this
information. Suppress the SELinux denials.

Change-Id: Iec4a997f0440ec793e2ebe919be017bba8b9a86c