Update patch set 4
Patch Set 4:

We may need to enable this globally when the time comes. Since we're currently working on a common gestures screen in CMParts, it may need to be placed in the AOSP gestures screen. It's not ready yet though, so I guess making this device-specific is fine for now.

Patch-set: 4
Reviewer: Zhao Wei Liew <7880@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0